@@ -1,6 +1,6 @@
 module github.com/decred/dcrd/addrmgr/v2
 
-go 1.11
+go 1.16
 
 require (
 	github.com/decred/dcrd/chaincfg/chainhash v1.0.2
