@@ -1,6 +1,6 @@
 module github.com/decred/dcrd
 
-go 1.13
+go 1.16
 
 require (
 	github.com/davecgh/go-spew v1.1.1
