@@ -10,7 +10,6 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"testing"
@@ -70,7 +69,7 @@ func createTestDatabase(dbName string, dbType string, net wire.CurrencyNet) (dat
 		}
 	} else {
 		// Create the directory for the test database.
-		dbPath, err := ioutil.TempDir("", dbName)
+		dbPath, err := os.MkdirTemp("", dbName)
 		if err != nil {
 			err := fmt.Errorf("unable to create test db path: %w",
 				err)
