@@ -14,7 +14,6 @@ import (
 	"encoding/binary"
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"math"
 	"net"
 	"os"
@@ -3205,10 +3204,10 @@ func genCertPair(certFile, keyFile string, altDNSNames []string, tlsCurve ellipt
 	}
 
 	// Write cert and key files.
-	if err = ioutil.WriteFile(certFile, cert, 0644); err != nil {
+	if err = os.WriteFile(certFile, cert, 0644); err != nil {
 		return err
 	}
-	if err = ioutil.WriteFile(keyFile, key, 0600); err != nil {
+	if err = os.WriteFile(keyFile, key, 0600); err != nil {
 		os.Remove(certFile)
 		return err
 	}
@@ -3255,7 +3254,7 @@ func setupRPCListeners() ([]net.Listener, error) {
 		}
 
 		if cfg.RPCAuthType == authTypeClientCert {
-			pemCerts, err := ioutil.ReadFile(cfg.RPCClientCAs)
+			pemCerts, err := os.ReadFile(cfg.RPCClientCAs)
 			if err != nil {
 				return nil, err
 			}
